flask_app/requirements.txt
Flask
flask-swagger-ui
Flask-Migrate
paho-mqtt==1.6.1
Flask-SQLAlchemy